requirements.txt
pillow>=11.0.0
numpy>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0